% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCarbonPrice.R
\name{calcCarbonPrice}
\alias{calcCarbonPrice}
\title{Calculate Carbon Price value and emisssions coverage based on the World Bank dashboard data}
\usage{
calcCarbonPrice(subtype = "effectivePrice")
}
\arguments{
\item{subtype}{data subtype. Either "emissionsCovered", "shareEmissionsCovered", "carbonPrice", "effectivePrice"}
}
\value{
MAgPIE object with Carbon Price data per country and sector group
}
\description{
Carbon Price data is calculated from World Bank carbon price, World Bank carbon price coverage shares,
and EDGAR Global GHG Emissions for sectoral disaggregation.
}
\examples{
\dontrun{
calcOutput("CarbonPrice", subtype = "effectivePrice")
}

}
\author{
Renato Rodrigues
}
