% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEffectiveCarbonPrice.R
\name{calcEffectiveCarbonPrice}
\alias{calcEffectiveCarbonPrice}
\title{Calculate the Effective Carbon Price based on the World Bank dashboard data}
\usage{
calcEffectiveCarbonPrice(subtype)
}
\arguments{
\item{subtype}{data subtype. Either "emissionsCovered", "shareEmissionsCovered", "carbonPrice", "effectivePrice"}
}
\value{
MAgPIE object with Effective Carbon Price per country and sector group
}
\description{
Effective Carbon Price is calculated from world bank emissions share Coverage,
EDGAR Global GHG Emissions and World Bank Carbon Price data.
}
\examples{
\dontrun{
calcOutput("WBCarbonPricingDashboard", subtype = "price")
}

}
\author{
Renato Rodrigues
}
