% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGlobalEconomyDataIndicators.R
\name{calcGlobalEconomyDataIndicators}
\alias{calcGlobalEconomyDataIndicators}
\title{Calculate the Global Economy Data indicators including assumptions to fill all missing country data}
\usage{
calcGlobalEconomyDataIndicators(outPeriod = 2003:2021)
}
\arguments{
\item{outPeriod}{vector of years to include in the output. Default is 2003:2021.
Make sure to include years with enough data.}
}
\value{
MAgPIE object with Global Economy Data indicators per indicator type, indicator name, indicator unit,
country and period
}
\description{
Calculate the Global Economy Data indicators including assumptions to fill all missing country data
}
\examples{
\dontrun{
calcOutput("GlobalEconomyDataIndicators", outPeriod = 2003:2021)
}

}
\author{
Renato Rodrigues
}
