% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGlobalEconomyData.R
\name{convertGlobalEconomyData}
\alias{convertGlobalEconomyData}
\title{Convert Global economy data indicators
aggregation of over 500 indicators for over 200 countries from 1960 - 2024.
sources: central banks, national statistical institutes, and multiple international organizations.
https://www.theglobaleconomy.com/}
\usage{
convertGlobalEconomyData(x)
}
\arguments{
\item{x}{A \code{\link[magclass:magclass-package]{magpie}} object returned from
\code{\link[=readWBCarbonPricingDashboard]{readWBCarbonPricingDashboard()}}.}
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Convert Global economy data indicators
aggregation of over 500 indicators for over 200 countries from 1960 - 2024.
sources: central banks, national statistical institutes, and multiple international organizations.
https://www.theglobaleconomy.com/
}
\author{
Renato Rodrigues
}
