% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertWBCarbonPricingDashboard.R
\name{convertWBCarbonPricingDashboard}
\alias{convertWBCarbonPricingDashboard}
\title{Convert Read World Bank Carbon Pricing Dashboard data}
\usage{
convertWBCarbonPricingDashboard(x, subtype)
}
\arguments{
\item{x}{A \code{\link[magclass:magclass-package]{magpie}} object returned from
\code{\link[=readWBCarbonPricingDashboard]{readWBCarbonPricingDashboard()}}.}

\item{subtype}{character, type of data: price, priceAprilFirst, wbCoverage, revenue, emissionsCovered}
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Convert Read World Bank Carbon Pricing Dashboard data
}
\author{
Renato Rodrigues
}
