% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrPEM-package.R
\docType{package}
\name{mrPEM-package}
\alias{mrPEM}
\alias{mrPEM-package}
\title{mrPEM: Prepare data to be used by the Political Economy Model (PEM)}
\description{
Prepare data to be used by the Political Economy Model (PEM).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/mrPEM}
}

}
\author{
\strong{Maintainer}: Renato Rodrigues \email{renato.rodrigues@pik-potsdam.de} (\href{https://orcid.org/0000-0002-5863-5514}{ORCID})

Authors:
\itemize{
  \item Kriegler Elmar \email{kriegler@pik-potsdam.de} (\href{https://orcid.org/0000-0002-3307-2647}{ORCID})
}

}
\keyword{internal}
