% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGlobalEconomyData.R
\name{readGlobalEconomyData}
\alias{readGlobalEconomyData}
\title{Global economy data indicators
aggregation of over 500 indicators for over 200 countries from 1960 - 2024.
sources: central banks, national statistical institutes, and multiple international organizations.
https://www.theglobaleconomy.com/}
\usage{
readGlobalEconomyData()
}
\value{
MagPIE object with global economy data indicators
}
\description{
Global economy data indicators
aggregation of over 500 indicators for over 200 countries from 1960 - 2024.
sources: central banks, national statistical institutes, and multiple international organizations.
https://www.theglobaleconomy.com/
}
\author{
Renato Rodrigues
}
