% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWBCarbonPricingDashboard.R
\name{readWBCarbonPricingDashboard}
\alias{readWBCarbonPricingDashboard}
\title{Read World Bank Carbon Pricing Dashboard data}
\usage{
readWBCarbonPricingDashboard(subtype = "price")
}
\arguments{
\item{subtype}{character, type of data: price, price_1_april, coverage, revenue, emissions_share or emissions}
}
\value{
MagPIE object with world banck data on carbon price per country
}
\description{
This data can be downloaded directly from the site.
}
\author{
Renato Rodrigues
}
