% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEffectiveCarbonPrice.R
\name{calcEffectiveCarbonPrice}
\alias{calcEffectiveCarbonPrice}
\title{calc WBCarbonPricingDashboard}
\usage{
calcEffectiveCarbonPrice()
}
\value{
MAgPIE object with Effective Carbon Price per country and sector group
}
\description{
calculate the Effective Carbon Price based on the World Bank dashboard data
}
\note{
Effective Carbon Price is calculated from world banck emissions share coverage,
EDGAR Global GHG Emissions and World Bank Carbon Price data.
}
\examples{
\dontrun{
calcOutput("WBCarbonPricingDashboard", subtype = "price")
}

}
\author{
Renato Rodrigues
}
