% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGlobalEconomyDataIndicators.R
\name{calcGlobalEconomyDataIndicators}
\alias{calcGlobalEconomyDataIndicators}
\title{Calculate the Effective Carbon Price based on the World Bank dashboard data}
\usage{
calcGlobalEconomyDataIndicators(outPeriod = 2003:2021)
}
\arguments{
\item{outPeriod}{vector of years to include in the output}
}
\value{
MAgPIE object with Effective Carbon Price per country and sector group
}
\description{
Effective Carbon Price is calculated from world bank emissions share Coverage,
EDGAR Global GHG Emissions and World Bank Carbon Price data.
}
\examples{
\dontrun{
calcOutput("GlobalEconomyDataIndicators", outPeriod = 2003:2021)
}

}
\author{
Renato Rodrigues
}
